Control de accesos y administracion de telefonos

+++ regiones/delete.php
@@ -8,39 +8,45 @@
 
     session_start();
 
-    if (isset($_POST['confirm']) && $_POST['confirm'] == 1) {
-        $region_id = (int) $_POST['region'];
-        #consultar por la region que se quiere eliminar
-        $res = $mbd->prepare("SELECT id FROM regiones WHERE id = ?");
-        $res->bindParam(1, $region_id);
-        $res->execute();
-        $region = $res->fetch();
+    if(isset($_SESSION['autenticado']) && $_SESSION['usuario_rol'] == 'Administrador(a)'){
 
-        if ($region) {
-            # verificar que la region no tenga comunas asociadas
-            $res = $mbd->prepare("SELECT id FROM comunas WHERE region_id = ?");
+        if (isset($_POST['confirm']) && $_POST['confirm'] == 1) {
+            $region_id = (int) $_POST['region'];
+            #consultar por la region que se quiere eliminar
+            $res = $mbd->prepare("SELECT id FROM regiones WHERE id = ?");
             $res->bindParam(1, $region_id);
             $res->execute();
-            $comunas = $res->fetchall();
+            $region = $res->fetch();
 
-            if (!count($comunas)) {
-                #eliminar la region
-                $res = $mbd->prepare("DELETE FROM regiones WHERE id = ?");
+            if ($region) {
+                # verificar que la region no tenga comunas asociadas
+                $res = $mbd->prepare("SELECT id FROM comunas WHERE region_id = ?");
                 $res->bindParam(1, $region_id);
                 $res->execute();
+                $comunas = $res->fetchall();
 
-                $row = $res->rowCount();
+                if (!count($comunas)) {
+                    #eliminar la region
+                    $res = $mbd->prepare("DELETE FROM regiones WHERE id = ?");
+                    $res->bindParam(1, $region_id);
+                    $res->execute();
 
-                if ($row) {
-                    $_SESSION['success'] = 'La región se ha eliminado correctamente';
-                    header('Location: ' . REGIONES);
+                    $row = $res->rowCount();
+
+                    if ($row) {
+                        $_SESSION['success'] = 'La región se ha eliminado correctamente';
+                        header('Location: ' . REGIONES);
+                    }
+                }else {
+                    $_SESSION['danger'] = 'Esta región no puede eliminarse porque tiene comunas asociadas';
+                    header('Location: ' . REGIONES . 'show.php?id=' . $region_id);
                 }
             }else {
-                $_SESSION['danger'] = 'Esta región no puede eliminarse porque tiene comunas asociadas';
+                $_SESSION['danger'] = 'Esta región no existe...';
                 header('Location: ' . REGIONES . 'show.php?id=' . $region_id);
             }
-        }else {
-            $_SESSION['danger'] = 'Esta región no existe...';
-            header('Location: ' . REGIONES . 'show.php?id=' . $region_id);
         }
+    }else{
+        $_SESSION['danger'] = 'Operación no permitida';
+        header('Location: ' . BASE_URL);
     }
